Adds Search Functionality + Description
CraigslistQuery is now able to take a hash of search params which
correspond to the search params used by Craigslist to refine a search.
It will also ignore duplicate posts and the `nearby results` returned
when there are less than 120 results per page.

CraigslistPost now has access to the description of a post. Had to
remove a redundant div on the description which seems to hold some kind
of qrcode.