Replace torify with socksify
Instead of relying on calling torify as a system command directly, we
are now using Socksify to set up a TOR proxy.

`generate_new_tor_circuit` makes use of `net-telnet` gem in order to
establish a telnet connection to the tor port 9051. This is needed in
order to authenticate and then generate a new circuit that will give us
a new ip address.

This is nice because attempts to generate a new circuit by executing a
system command from within ruby was being blocked due to not being authenticated.
The direct telnet connection solves this problem.

Also, we are now resetting circuits (ip) when one gets blocked.