[Fix]: Duplicate Posts will no longer be saved
The `new_post` methods was not properly stopping duplicate posts from
being created.

We now have made `post_id` unique, with a db constraint and a model
validation. We are using this field to check incoming new
posts to see if we have already stored them. If we have we skip them.