[Fix]: Display time as it was stored
Rails will 'coerce' date into UTC when stored in db. In this case the
coercion does not work because the provided time from CL does not
include a time zone. So we do not want to do any additional time zone
conversion. We simply need to display the time we originally received.
This time will be in the timezome CL used.

It appears CL is using the ip location to determine the timezone to
return. The use of proxies will probably make the dates returned
unreliable.