Adds Favorites to CraigslistPosts
Users can now favorite individual posts. This will cause those posts to
always be displayed at the top of the posts table. Sorting will apply to
them just as it will to the rest of the table, they will simply be
grouped at the top.

Also, we are now persisting sorts when using a ransack search field.
Previously, sorts were cleared and default sorting was applied.