Fix: Nested Forms
HTML does not allow nested forms. I had nested a `search_form` from
ransack inside a `form_tag` and this was causing issues submitting the
search form.

The solution is to unnest forms by creating an empty form with an `id`
and give the form elements (inputs ans submit buttons) that id.