Allows User to Edit Alerts
When an Alert is edited, the existing `PullPostsJob` for that Alert will
be destroyed and a new one will be enqueued with the updated Alert. This
will ensure that any subsequent pulls will use the updated Alert fields.

Note on `SearchParams`: I am am using method_missing design pattern here
to avoid having a dozen individual methods defined to return their
corresponding `search_param`. In the future if the need to search the
database for Alerts based on the `search_params` arises, then we can add
databse fields and remove the method_missing.