Adds Ability to Search for Cities Available on CL
The `CraigslistSites` class will pull the list of all cities by states
that CL has available, along with their urls. This list is used to
populate the option for the `city` field on the Alerts new/edit form.

We are making use of `bootstrap-select` for the search functionality.
Due to the nature of turbolinks not playing well with JS, we had to
disable it on links to this form that contains the bootstrap select.